+++ src/components/link/demo/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var demoApp = angular.module('demoApp', [
+  'officeuifabric.core',
+  'officeuifabric.components.link'
+]);
+